#435 update ket schema
- remove format folder from ketcher-react
- create KetSerializer
- update 'text' schema
- add alert if cannot open file
