+++ packages/ketcher-core/src/chemistry/serializers/ket/fromGraph/rgroupToStruct.ts
@@ -1,11 +1,11 @@
-import { ifDef } from '../../utils'
+import { ifDef } from 'utils'
 import { moleculeToStruct } from './moleculeToStruct'
-import { RGroup } from 'ketcher-core'
+import { RGroup, Struct } from 'chemistry/entities'
 
-export function rgroupToStruct(graphItem) {
+export function rgroupToStruct(graphItem): Struct {
   const struct = moleculeToStruct(graphItem)
   const rgroup = rgroupLogicToStruct(graphItem.rlogic)
-  struct.frags.forEach((value, key) => {
+  struct.frags.forEach((_value: any, key) => {
     rgroup.frags.add(key)
   })
   if (graphItem.rlogic) struct.rgroups.set(graphItem.rlogic.number, rgroup)
