+++ packages/ketcher-core/src/chemistry/serializers/ket/index.ts
@@ -1,1 +1,1 @@
-export * from './ket.types'
+export * from './KetSerializer'
