+++ packages/ketcher-core/src/chemistry/serializers/ket/toGraph/textToGraph.ts
