+++ packages/ketcher-core/src/chemistry/serializers/ket/validate.ts
@@ -1,0 +1,23 @@
+/****************************************************************************
+ * Copyright 2020 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ ***************************************************************************/
+import { Validator } from 'jsonschema'
+import schema from './schema.json'
+
+export function validate(graph: any): boolean {
+  const validator = new Validator()
+  const result = validator.validate(graph, schema)
+  return result.valid
+}
