+++ packages/ketcher-core/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './Pile'
 export * from './Pool'
 export * from './scale'
 export * from './Vec2'
+export * from './ifDef'
