--- packages/ketcher-react/src/script/format/chemGraph/toGraph/rxnToGraph.js
@@ -1,18 +1,0 @@
-export function arrowToGraph(arrowNode) {
-  const coord = arrowNode.center
-  return {
-    type: 'arrow',
-    location: [coord.x, coord.y, coord.z],
-    prop: arrowNode.data,
-    mode: arrowNode.mode
-  }
-}
-
-export function plusToGraph(plusNode) {
-  const coord = plusNode.center
-  return {
-    type: 'plus',
-    location: [coord.x, coord.y, coord.z],
-    prop: plusNode.data
-  }
-}
